+++ lib/rules/no-useless-character-class.ts
@@ -35,9 +35,6 @@ export default createRule("no-useless-character-class", {
     create(context) {
         const ignores: string[] = context.options[0]?.ignores ?? ["="]
 
-        /**
-         * Create visitor
-         */
         function createVisitor({
             node,
             flags,
@@ -45,7 +42,6 @@ export default createRule("no-useless-character-class", {
             getRegexpLocation,
         }: RegExpContext): RegExpVisitor.Handlers {
             return {
-                // eslint-disable-next-line complexity -- X(
                 onCharacterClassEnter(ccNode) {
                     if (ccNode.elements.length !== 1) {
                         return