+++ lib/rules/no-useless-lazy.ts
@@ -58,9 +58,6 @@ export default createRule("no-useless-lazy", {
         type: "suggestion", // "problem",
     },
     create(context) {
-        /**
-         * Create visitor
-         */
         function createVisitor(
             regexpContext: RegExpContext,
         ): RegExpVisitor.Handlers {