+++ lib/rules/prefer-question-quantifier.ts
@@ -21,9 +21,6 @@ export default createRule("prefer-question-quantifier", {
         type: "suggestion", // "problem",
     },
     create(context) {
-        /**
-         * Create visitor
-         */
         function createVisitor({
             node,
             getRegexpLocation,