+++ lib/rules/prefer-w.ts
@@ -19,10 +19,6 @@ import {
 import { Chars, toUnicodeSet } from "regexp-ast-analysis"
 import { mention } from "../utils/mention"
 
-/**
- * Checks if small letter char class range
- * @param node The node to check
- */
 function isSmallLetterRange(node: CharacterClassElement) {
     return (
         node.type === "CharacterClassRange" &&
@@ -31,10 +27,6 @@ function isSmallLetterRange(node: CharacterClassElement) {
     )
 }
 
-/**
- * Checks if capital letter char class range
- * @param node The node to check
- */
 function isCapitalLetterRange(node: CharacterClassElement) {
     return (
         node.type === "CharacterClassRange" &&
@@ -43,10 +35,6 @@ function isCapitalLetterRange(node: CharacterClassElement) {
     )
 }
 
-/**
- * Checks if digit char class
- * @param node The node to check
- */
 function isDigitRangeOrSet(node: CharacterClassElement) {
     return (
         (node.type === "CharacterClassRange" &&
@@ -56,10 +44,6 @@ function isDigitRangeOrSet(node: CharacterClassElement) {
     )
 }
 
-/**
- * Checks if includes `_`
- * @param node The node to check
- */
 function isUnderscoreCharacter(node: CharacterClassElement) {
     return node.type === "Character" && node.value === CP_LOW_LINE
 }
@@ -80,9 +64,6 @@ export default createRule("prefer-w", {
         type: "suggestion", // "problem",
     },
     create(context) {
-        /**
-         * Create visitor
-         */
         function createVisitor({
             node,
             flags,