+++ lib/rules/unicode-escape.ts
@@ -34,9 +34,6 @@ export default createRule("unicode-escape", {
         const preferUnicodeCodePointEscape =
             context.options[0] !== "unicodeEscape"
 
-        /**
-         * Verify for unicodeCodePointEscape
-         */
         function verifyForUnicodeCodePointEscape(
             { node, getRegexpLocation, fixReplaceNode }: RegExpContext,
             kind: EscapeSequenceKind,
@@ -60,9 +57,6 @@ export default createRule("unicode-escape", {
             })
         }
 
-        /**
-         * Verify for unicodeEscape
-         */
         function verifyForUnicodeEscape(
             { node, getRegexpLocation, fixReplaceNode }: RegExpContext,
             kind: EscapeSequenceKind,
@@ -90,9 +84,6 @@ export default createRule("unicode-escape", {
             ? verifyForUnicodeCodePointEscape
             : verifyForUnicodeEscape
 
-        /**
-         * Create visitor
-         */
         function createVisitor(
             regexpContext: RegExpContext,
         ): RegExpVisitor.Handlers {