+++ lib/utils/regexp-ast/is-covered.ts
@@ -335,10 +335,8 @@ export function isCoveredNode(
     return isCoveredForNormalizedNode(leftNode, rightNode, options)
 }
 
-/* eslint-disable complexity -- X( */
 /** Checks whether the right node is covered by the left node. */
 function isCoveredForNormalizedNode(
-    /* eslint-enable complexity -- X( */
     left: NormalizedNode,
     right: NormalizedNode,
     options: Options,
@@ -412,7 +410,6 @@ function isCoveredForNormalizedNode(
 
 const cacheNormalizeNode = new WeakMap<Node, NormalizedNode>()
 
-/** Normalize node */
 function normalizeNode(node: Node, options: Options): NormalizedNode {
     let n = cacheNormalizeNode.get(node)
     if (n) {
@@ -424,7 +421,6 @@ function normalizeNode(node: Node, options: Options): NormalizedNode {
     return n
 }
 
-/** Normalize node without cache */
 function normalizeNodeWithoutCache(
     node: Node,
     options: Options,
@@ -478,10 +474,8 @@ function isCoveredAnyNode(
     return false
 }
 
-/* eslint-disable complexity -- X( */
 /** Check whether the right nodes is covered by the left nodes. */
 function isCoveredAltNodes(
-    /* eslint-enable complexity -- X( */
     leftNodes: NormalizedNode[],
     rightNodes: NormalizedNode[],
     options: Options,